Раскидал часть файлов и переделал плеер
