+++ app/src/main/java/com/example/playlistmaker/App.kt
@@ -1,0 +1,42 @@
+package com.example.playlistmaker
+
+import android.app.Application
+import android.util.Log
+import androidx.appcompat.app.AppCompatDelegate
+import com.example.playlistmaker.sharing.data.ExternalNavigator
+
+const val HISTORY_COUNT: Int = 10
+const val PLAY_LIST_PREFERENCES = "play_list_preferences"
+
+class App : Application() {
+    private lateinit var externalNavigator: ExternalNavigator
+    var darkTheme = false
+    override fun onCreate() {
+        super.onCreate()
+
+        Log.println(Log.INFO, "my_tag", "onCreate.App")
+        externalNavigator = ExternalNavigator(getSharedPreferences(PLAY_LIST_PREFERENCES, MODE_PRIVATE))
+        initApplicationPreferences()
+    }
+
+    private fun initApplicationPreferences() {
+        darkTheme = externalNavigator.restoreTheme()
+        switchTheme()
+    }
+
+    fun switchThemeAndSave(darkThemeEnabled: Boolean) {
+        darkTheme = darkThemeEnabled
+        switchTheme()
+        externalNavigator.saveTheme(darkThemeEnabled)
+    }
+
+    private fun switchTheme() {
+        AppCompatDelegate.setDefaultNightMode(
+            if (darkTheme) {
+                AppCompatDelegate.MODE_NIGHT_YES
+            } else {
+                AppCompatDelegate.MODE_NIGHT_NO
+            }
+        )
+    }
+}
