--- app/src/main/java/com/example/playlistmaker/data/network/ItunesApiService.kt
@@ -1,11 +1,0 @@
-package com.example.playlistmaker.data.network
-
-import com.example.playlistmaker.data.dto.TracksSearchResponse
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface ItunesApiService {
-    @GET("/search?entity=song")
-    fun searchTracks(@Query("term") text: String): Call<TracksSearchResponse>
-}
