--- app/src/main/java/com/example/playlistmaker/data/network/TracksApi.kt
@@ -1,11 +1,0 @@
-package com.example.playlistmaker.data.network
-
-import com.example.playlistmaker.domain.models.track.Tracks
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface TracksApi {
-    @GET("/search?entity=song")
-    fun searchTracks(@Query("term") text: String): Call<Tracks>
-}
