--- app/src/main/java/com/example/playlistmaker/domain/api/PlayerInteractor.kt
@@ -1,19 +1,0 @@
-package com.example.playlistmaker.domain.api
-
-import com.example.playlistmaker.domain.player.PlayerState
-
-interface PlayerInteractor {
-    fun preparePlayer(streamUrl: String)
-
-    fun setPlayerState(state: PlayerState)
-
-    fun getPlayerState(): PlayerState
-
-    fun getCurrentPosition(): Int
-
-    fun pausePlayer()
-
-    fun startPlayer()
-
-    fun releasePlayer()
-}
