--- app/src/main/java/com/example/playlistmaker/domain/api/TrackRepository.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker.domain.api
-
-import com.example.playlistmaker.domain.models.track.Track
-
-interface TrackRepository {
-    fun searchTracks(expression: String): MutableList<Track>
-}
