--- app/src/main/java/com/example/playlistmaker/domain/player/PlayerRepository.kt
@@ -1,18 +1,0 @@
-package com.example.playlistmaker.domain.player
-
-interface PlayerRepository {
-
-    fun preparePlayer(streamUrl: String)
-
-    fun setPlayerState(state: PlayerState)
-
-    fun getPlayerState(): PlayerState
-
-    fun getCurrentPosition(): Int
-
-    fun pausePlayer()
-
-    fun startPlayer()
-
-    fun releasePlayer()
-}
