--- app/src/main/java/com/example/playlistmaker/domain/player/PlayerState.kt
@@ -1,8 +1,0 @@
-package com.example.playlistmaker.domain.player
-
-enum class PlayerState {
-    STATE_DEFAULT,
-    STATE_PREPARED,
-    STATE_PLAYING,
-    STATE_PAUSED
-}
