+++ app/src/main/java/com/example/playlistmaker/player/data/GlideRepository.kt
@@ -1,0 +1,7 @@
+//package com.example.playlistmaker.player.data
+//
+//import android.widget.ImageView
+//
+//interface GlideRepository {
+//    fun glideBind(link: String, target: ImageView, placeholderId: Int)
+//}
