+++ app/src/main/java/com/example/playlistmaker/player/domain/GlideInteractor.kt
@@ -1,0 +1,7 @@
+//package com.example.playlistmaker.player.domain
+//
+//import android.widget.ImageView
+//
+//interface GlideInteractor {
+//    fun glideBind(link: String, target: ImageView, placeholderId: Int)
+//}
