+++ app/src/main/java/com/example/playlistmaker/player/domain/GlideInteractorImpl.kt
@@ -1,0 +1,11 @@
+//package com.example.playlistmaker.player.domain
+//
+//import android.widget.ImageView
+//import com.example.playlistmaker.player.data.GlideRepository
+//
+//class GlideInteractorImpl(private val repository: GlideRepository) : GlideInteractor {
+//    override fun glideBind(link: String, target: ImageView, placeholderId: Int) {
+//        repository.glideBind(link, target, placeholderId)
+//    }
+//
+//}
