+++ app/src/main/java/com/example/playlistmaker/player/domain/PlayerState.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.player.domain
+
+enum class PlayerState {
+    STATE_DEFAULT,
+    STATE_PREPARED,
+    STATE_PLAYING,
+    STATE_PAUSED
+}
