+++ app/src/main/java/com/example/playlistmaker/player/domain/TrackActivity.kt
@@ -1,0 +1,14 @@
+//package com.example.playlistmaker.player.domain
+//
+//import android.os.Bundle
+//import androidx.activity.ComponentActivity
+//import androidx.lifecycle.ViewModelProvider
+//
+//class TrackActivity : ComponentActivity() {
+//    private lateinit var viewModel: PlayerViewModel
+//
+//    override fun onCreate(savedInstanceState: Bundle?) {
+//        super.onCreate(savedInstanceState)
+//        viewModel = ViewModelProvider(this)[PlayerViewModel::class.java]
+//    }
+//}
