+++ app/src/main/java/com/example/playlistmaker/search/data/TrackRepository.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.search.data
+
+import com.example.playlistmaker.domain.models.track.Track
+
+interface TrackRepository {
+    fun searchTracks(expression: String): MutableList<Track>
+}
