+++ app/src/main/java/com/example/playlistmaker/search/recyclerview/SearchAdapter.kt
@@ -1,0 +1,55 @@
+package com.example.playlistmaker.search.recyclerview
+
+import android.os.Handler
+import android.os.Looper
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import com.example.playlistmaker.R
+import com.example.playlistmaker.domain.models.track.Track
+
+class SearchAdapter (private val data: List<Track>) : RecyclerView.Adapter<SearchViewHolder> (){
+    companion object {
+        private const val CLICK_DEBOUNCE_DELAY = 1000L
+    }
+
+    private var isClickAllowed = true
+    private val handler = Handler(Looper.getMainLooper())
+    private var onClickListener: OnClickListener? = null
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): SearchViewHolder {
+        val view = LayoutInflater.from(parent.context).inflate(R.layout.search_item, parent, false)
+        return SearchViewHolder(view)
+    }
+
+    override fun onBindViewHolder(holder: SearchViewHolder, position: Int) {
+        val item: Track = data[position]
+        holder.bind(item)
+        holder.itemView.setOnClickListener {
+            if (onClickListener != null && clickDebounce()){
+                onClickListener!!.onClick(position, item)
+            }
+        }
+    }
+
+    override fun getItemCount(): Int {
+        return data.size
+    }
+
+    fun setOnClickListener(onClickListener: OnClickListener){
+        this.onClickListener = onClickListener
+    }
+
+    interface  OnClickListener{
+        fun onClick(position: Int, track: Track)
+    }
+
+    private fun clickDebounce() : Boolean {
+        val current = isClickAllowed
+        if (isClickAllowed) {
+            isClickAllowed = false
+            handler.postDelayed({ isClickAllowed = true }, CLICK_DEBOUNCE_DELAY)
+        }
+        return current
+    }
+}
