+++ app/src/main/java/com/example/playlistmaker/settings/ui/SettingsViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.playlistmaker.settings.ui
+
+class SettingsViewModel {
+}
